pyproject.toml
[tool.poetry]
name = "occasion_back"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
django-enumfields = "^2.1.1"
Pyrogram = "^2.0.27"
psycopg2 = "^2.9.3"
TgCrypto = "^1.2.3"
celery = "^5.2.7"
django-cryptography = "^1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
